+++ database/migrations/2022_02_18_013750_create_operations_types_table.php
@@ -11,12 +11,12 @@ class CreateOperationsTypesTable extends Migration
         Schema::create('operations_types', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->bigInteger('reference_sequence_id')->nullable();
+            $table->unsignedInteger('reference_sequence_id')->nullable();
             $table->string('code');
-            $table->bigInteger('warehouse_id')->nullable();
+            $table->unsignedInteger('warehouse_id')->nullable();
             $table->string('reservation_method')->nullable();
             $table->string('type');
-            $table->bigInteger('operation_type_for_returns_id')->nullable();
+            $table->unsignedInteger('operation_type_for_returns_id')->nullable();
             $table->boolean('show_detailed_operation')->default(false);
             $table->boolean('pre_fill_detailed_operation')->default(false);
             $table->unsignedInteger('reservation_days_before')->nullable()->default(0);
@@ -24,8 +24,8 @@ class CreateOperationsTypesTable extends Migration
             $table->boolean('create_new_lots_serial_numbers')->default(false);
             $table->boolean('use_existing_lots_serial_numbers')->default(false);
             $table->boolean('create_new_lots_serial_numbers_for_components')->default(false); // only shows at manufacturing type
-            $table->bigInteger('default_source_location_id')->nullable();
-            $table->bigInteger('default_destination_location_id')->nullable();
+            $table->unsignedInteger('default_source_location_id')->nullable();
+            $table->unsignedInteger('default_destination_location_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
