+++ database/migrations/2022_03_10_014850_create_app_menus_table.php
@@ -11,8 +11,8 @@ class CreateAppMenusTable extends Migration
         Schema::create('app_menus', function (Blueprint $table) {
             $table->id();
             $table->string('label');
-            $table->bigInteger('menu_id')->nullable();
-            $table->bigInteger('parent_app_menu_id')->nullable();
+            $table->unsignedInteger('menu_id')->nullable();
+            $table->unsignedInteger('parent_app_menu_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
