+++ database/migrations/2022_03_28_235608_create_adjustment_lines_table.php
@@ -10,9 +10,9 @@ class CreateAdjustmentLinesTable extends Migration
     {
         Schema::create('adjustment_lines', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('adjustment_id')->nullable();
-            $table->bigInteger('product_id')->nullable();
-            $table->bigInteger('measurement_id');
+            $table->unsignedInteger('adjustment_id')->nullable();
+            $table->unsignedInteger('product_id')->nullable();
+            $table->unsignedInteger('measurement_id');
             $table->integer('quantity_on_hand');
             $table->integer('quantity_counted');
             $table->softDeletes();
