+++ database/migrations/2022_05_12_012019_create_delivery_fees_table.php
@@ -12,7 +12,7 @@ class CreateDeliveryFeesTable extends Migration
             $table->id();
             $table->string('name')->nullable();
             $table->boolean('is_enabled')->default(false);
-            $table->bigInteger('product_id')->nullable();
+            $table->unsignedInteger('product_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
